launch scripts override GAZEBO_MASTER_URI
GAZEBO_MASTER_URI is always the one written in `setup.sh`, even if a different value is set in the user's environment.

I consider it counter-intuitive, if you e.g. run `GAZEBO_MASTER_URI=http://myserver:11345 rosrun gazebo_ros gzserver`, that the server is still started with the default URI.