Adding ros1_bridge mapping rules for ODEJointStates message.
The hiStop and loStop fields in this message were renamed to
hi_stop and lo_stop to comply with ROS 2 naming rules.

This prevented the automatic mapping rules in ros1_bridge from
generating mapping code.

This change adds a manual mapping to allow the ros1_bridge to
successfully generate mappings for this message.