Made the MultiCameraPlugin really multi-camera.
Backwards compatibility should be retained.

SDF parameter <frameName> became optionally a comma-separated list specifying a different frame for each camera.

Optional SDF parameter <cameraSuffixes> was added. If empty or non-present, "/<sdf_camera_name>" is used as the suffix. If nonempty, it is treated as a comma-separated list of camera suffixes. This list has to have the same lengths as the number of cameras. Any separators (like "/") should be part of the specified suffixes.