gazebo_ros_control: catch all pluginlib exceptions
Before we only displayed an error message on LibraryLoadExceptions,
but there is also ClassLoaderException, CreateClassException,
InvalidXMLException, and some more. So catch
pluginlib::PluginlibException, which is base class of all of these.

Uncaught exceptions don't show up at all (probably a bug in gazebo itself),
which leads to situations which are quite hard to debug.