WheelSlip.cfg: ignore default parameter values
The WheelSlipPlugin already has default values for the slip
compliance parameters, so ignore the default values set in
WheelSlip.cfg by making them negative and instruct the
callback to ignore negative values.
Also, ignore the `level` variable in callback so that
parameters set in launch file will not be ignored.