Support parameter value override in launch file
Similar to the change in #1111, moves the WheelSlipPlugin::Load
call before the parameter callback so that the values from SDF/URDF
are set first.
Then the callback is changed to ignore negative values, and the
default slip parameter values are set to -1, so that the SDF/URDF
values are still preferred unless a different parameter value
is specified in a launch file.