Subscribe to gazebo topic on ROS topic connection
We are currently subscribing to the /gazebo/performance_metrics topic
even if there are no subscribers to the ROS topic forwarding this data.
The link_states and model_states topics currently use an advertise
mechanism with callbacks when a subscriber connects or disconnects,
so I've used that same pattern for the performance_metrics topic.
This also helps workaround the deadlock documented in #1175 and
osrf/gazebo#2902.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>