Fix use of mutexes
Restore original scope for wait_mutex_.
Use a separate set of mutexes for making adding/removing nodes thread-safe.
The set of mutexes is used to give priority to threads doing the adding/removing so that they are not constantly blocked by the quick spin threads.

This eliminates the need for the timeout in the spin threads.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>