Deprecate -spawn_service_timeout option
The spawn_entity.py script has two options for setting service timeouts:

1. '-timeout': Sets the duration to wait for all services to be ready
2. '-spawn_service_timeout': Sets the duration to wait for the spawn_entity service to be ready

(1) is available in all ROS distros since Dashing (except Eloquent for some reason).
(2) was introduced in Eloquent, and replaced (1) in Foxy as the way to control the timeout specifically for the spawn_entity service.

This change deprecates (2) and assumes (1) is used for all service timeouts again.

This change also adds smoke tests for the two options.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>