Fix console error when launching gzserver
If one of the substitutions in the gzserver command results in an empty string,
then we see the following console error and warning:

    [Err] [Server.cc:472] Could not open file[]
    [Wrn] [Server.cc:381] Falling back on worlds/empty.world

This is because gzserver is parsing any empty strings as arguments.
While we can probably fix this in launch, as a workaround we can filter out
any empty strings by using an OpaqueFunction to first process the substitutions
before returning the ExecuteProcess action to start gzserver.

It doesn't seem like we need a similar change for gzclient, since it does not produce any warnings or errors when given empty string arguments.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>