Fix console error when launching gzserver
If one of the substitutions in the gzserver command results in an empty string, then we see the following warning:

    [Err] [Server.cc:472] Could not open file[]
    [Wrn] [Server.cc:381] Falling back on worlds/empty.world

This is because gzserver is parsing any empty string as the world argument.
To fix the issue, we explicitly launch an empty world if no launch argument for the
world is given.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>